Fix MacOSX build script and CMake files for VCPKG

+++ build_osx.sh
@@ -9,26 +9,36 @@ git checkout tags/2022.08.15
 ./bootstrap-vcpkg.sh
 cd ..
 
-./vcpkg/vcpkg install libiconv:x64-osx
-./vcpkg/vcpkg install zlib:x64-osx
-./vcpkg/vcpkg install freetype:x64-osx
-./vcpkg/vcpkg install podofo:x64-osx
-./vcpkg/vcpkg install libxml2:x64-osx
-./vcpkg/vcpkg install leptonica:x64-osx
-./vcpkg/vcpkg install tesseract:x64-osx
-./vcpkg/vcpkg install boost-filesystem:x64-osx
-./vcpkg/vcpkg install boost-system:x64-osx
-./vcpkg/vcpkg install boost-signals2:x64-osx
-./vcpkg/vcpkg install boost-config:x64-osx
-./vcpkg/vcpkg install boost-dll:x64-osx
-./vcpkg/vcpkg install boost-assert:x64-osx
-./vcpkg/vcpkg install boost-smart-ptr:x64-osx
+VCPKG_TRIPLET=x64-osx
+
+./vcpkg/vcpkg install libiconv:$VCPKG_TRIPLET
+./vcpkg/vcpkg install zlib:$VCPKG_TRIPLET
+./vcpkg/vcpkg install freetype:$VCPKG_TRIPLET
+./vcpkg/vcpkg install libxml2:$VCPKG_TRIPLET
+./vcpkg/vcpkg install leptonica:$VCPKG_TRIPLET
+./vcpkg/vcpkg install tesseract:$VCPKG_TRIPLET
+./vcpkg/vcpkg install boost-filesystem:$VCPKG_TRIPLET
+./vcpkg/vcpkg install boost-system:$VCPKG_TRIPLET
+./vcpkg/vcpkg install boost-signals2:$VCPKG_TRIPLET
+./vcpkg/vcpkg install boost-config:$VCPKG_TRIPLET
+./vcpkg/vcpkg install boost-dll:$VCPKG_TRIPLET
+./vcpkg/vcpkg install boost-assert:$VCPKG_TRIPLET
+./vcpkg/vcpkg install boost-smart-ptr:$VCPKG_TRIPLET
+
+mkdir -p custom-triplets
+cp ./vcpkg/triplets/community/$VCPKG_TRIPLET-dynamic.cmake custom-triplets/$VCPKG_TRIPLET.cmake
+./vcpkg/vcpkg install podofo:$VCPKG_TRIPLET --overlay-triplets=custom-triplets
+
+vcpkg_prefix="$PWD/vcpkg/installed/$VCPKG_TRIPLET"
+
+deps_prefix="$PWD/deps"
+mkdir -p $deps_prefix
 
 wget -nc https://sourceforge.net/projects/htmlcxx/files/v0.87/htmlcxx-0.87.tar.gz
 echo "ac7b56357d6867f649e0f1f699d9a4f0f03a6e80  htmlcxx-0.87.tar.gz" | shasum -c
 tar -xzvf htmlcxx-0.87.tar.gz
 cd htmlcxx-0.87
-./configure CXXFLAGS=-std=c++17 LDFLAGS="-L$PWD/../vcpkg/packages/libiconv_x64-osx/lib" LIBS="-liconv" CPPFLAGS="-I$PWD/../vcpkg/packages/libiconv_x64/include"
+./configure CXXFLAGS=-std=c++17 LDFLAGS="-L$vcpkg_prefix/lib" LIBS="-liconv" CPPFLAGS="-I$vcpkg_prefix/include" --prefix="$deps_prefix"
 sed -i.bak -e "s/\(allow_undefined=\)yes/\1no/" libtool
 sed -i -r -e 's/css\/libcss_parser_pp.la \\//' Makefile
 sed -i -r -e 's/css\/libcss_parser.la//' Makefile
@@ -41,15 +51,16 @@ echo 'int main() {}' > htmlcxx.cc
 make -j4
 make install-strip
 cd ..
-
+rm -rf htmlcxx-0.87
 
 wget -nc http://silvercoders.com/download/3rdparty/libcharsetdetect-master.tar.bz2
 echo "6f9adaf7b6130bee6cfac179e3406cdb933bc83f  libcharsetdetect-master.tar.bz2" | shasum -c
 tar -xjvf libcharsetdetect-master.tar.bz2
 cd libcharsetdetect-master
-cmake -DCMAKE_CXX_STANDARD=17 -DBUILD_SHARED_LIBS=TRUE .
+cmake -DCMAKE_CXX_STANDARD=17 -DBUILD_SHARED_LIBS=TRUE -DCMAKE_INSTALL_PREFIX:PATH="$deps_prefix" .
 cmake --build . --config Release --target install
 cd ..
+rm -rf libcharsetdetect-master
 rm libcharsetdetect-master.tar.bz2
 
 
@@ -57,13 +68,13 @@ git clone https://github.com/docwire/wv2.git
 cd wv2
 mkdir build
 cd build
-cmake -DCMAKE_CXX_STANDARD=17 -DCMAKE_BUILD_TYPE=Debug ..
+cmake -DCMAKE_CXX_STANDARD=17 -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX:PATH="$deps_prefix" ..
 make install
 cd ..
 wget http://silvercoders.com/download/3rdparty/wv2-0.2.3_patched_4-private_headers.tar.bz2
 echo "6bb3959d975e483128623ee3bff3fba343f096c7  wv2-0.2.3_patched_4-private_headers.tar.bz2" | shasum -c
 tar -xjvf wv2-0.2.3_patched_4-private_headers.tar.bz2
-mv wv2-0.2.3_patched_4-private_headers/*.h /usr/local/include/wv2/
+mv wv2-0.2.3_patched_4-private_headers/*.h $deps_prefix/include/wv2/
 cd ..
 
 wget -nc http://www.codesink.org/download/mimetic-0.9.7.tar.gz
@@ -77,7 +88,7 @@ patch -p1 -i ../mimetic-0.9.7-patches/register_keyword.patch
 patch -p1 -i ../mimetic-0.9.7-patches/ContTokenizer.patch
 patch -p1 -i ../mimetic-0.9.7-patches/macro_string.patch
 patch -p1 -i ../mimetic-0.9.7-patches/mimetic_pointer_comparison.patch
-./configure CXXFLAGS=-std=c++17
+./configure CXXFLAGS=-std=c++17 --prefix="$deps_prefix"
 make -j4
 make install-strip
 cd ..
@@ -87,7 +98,7 @@ cd libbfio
 git checkout 3bb082c
 ./synclibs.sh
 autoreconf -i
-./configure
+./configure --prefix="$deps_prefix"
 make -j4
 make install
 cd ..
@@ -98,7 +109,7 @@ git checkout 99a86ef
 ./synclibs.sh
 touch ../../config.rpath
 autoreconf -i
-./configure
+./configure --prefix="$deps_prefix"
 make -j4
 make install
 cd ..
@@ -113,57 +124,45 @@ printf 'add_library(unzip STATIC ${UNZIP_SRC})\n' >> CMakeLists.txt
 printf 'install(FILES unzip.h ioapi.h DESTINATION include)\n' >> CMakeLists.txt
 printf 'install(TARGETS unzip DESTINATION lib)\n' >> CMakeLists.txt
 printf 'target_compile_options(unzip PRIVATE -fPIC)\n' >> CMakeLists.txt
-cmake -DCMAKE_CXX_STANDARD=17 .
+cmake -DCMAKE_CXX_STANDARD=17 -DCMAKE_INSTALL_PREFIX:PATH="$deps_prefix" -DCMAKE_CXX_FLAGS="-I$vcpkg_prefix/include" .
 cmake --build .
 cmake --install .
 cd ..
 
-vcpkg_cxx_flags=""
-for vcpkg_inc_dir in $PWD/vcpkg/packages/*/include; do vcpkg_cxx_flags="$vcpkg_cxx_flags -I$vcpkg_inc_dir"; done
-
-vcpkg_linker_flags=""
-for vcpkg_lib_dir in $PWD/vcpkg/packages/*/lib; do vcpkg_linker_flags="$vcpkg_linker_flags -L$vcpkg_lib_dir"; done
-
 mkdir -p build
 cd build
 cmake -DCMAKE_CXX_STANDARD=17 -DCMAKE_TOOLCHAIN_FILE=$PWD/../vcpkg/scripts/buildsystems/vcpkg.cmake \
-	-DCMAKE_CXX_FLAGS="$vcpkg_cxx_flags" \
-	-DCMAKE_SHARED_LINKER_FLAGS="$vcpkg_linker_flags" \
+	-DCMAKE_PREFIX_PATH="$deps_prefix" \
 	..
 cmake --build .
 cmake --build . --target doxygen install
 cd ..
 
-
-pwd
-cd build/
-cp -a /usr/local/share/tessdata/ tessdata/
-cp /usr/local/lib/libboost_filesystem.dylib .
-cp /usr/local/lib/libboost_system.dylib .
-cp /usr/local/lib/libwv2.4.dylib .
-cp /usr/local/opt/tesseract/lib/libtesseract.5.dylib .
-cp /usr/local/opt/podofo/lib/libpodofo.0.9.8.dylib .
-cp /usr/local/opt/freetype/lib/libfreetype.6.dylib .
-cp /usr/local/lib/libhtmlcxx.3.dylib .
-cp /usr/local/lib/libcharsetdetect.dylib .
-cp /usr/local/lib/libmimetic.0.dylib .
-cp /usr/local/lib/libbfio.1.dylib .
-cp /usr/local/lib/libpff.1.dylib .
-cp /usr/local/opt/libpng/lib/libpng16.16.dylib .
-cp /usr/local/opt/jpeg-turbo/lib/libjpeg.8.dylib .
-cp /usr/local/opt/giflib/lib/libgif.dylib .
-cp /usr/local/opt/libtiff/lib/libtiff.6.dylib .
-cp /usr/local/opt/webp/lib/libwebpmux.3.dylib .
-cp /usr/local/opt/webp/lib/libwebp.7.dylib .
-cp /usr/local/opt/openjpeg/lib/libopenjp2.7.dylib .
-ls
+cd build
+mkdir -p tessdata
+cd tessdata
+wget -nc https://github.com/tesseract-ocr/tessdata_fast/raw/main/eng.traineddata
+wget -nc https://github.com/tesseract-ocr/tessdata_fast/raw/main/osd.traineddata
+wget -nc https://github.com/tesseract-ocr/tessdata_fast/raw/main/pol.traineddata
+cd ..
+cp $deps_prefix/lib/libwv2.4.dylib .
+cp $vcpkg_prefix/lib/libpodofo.0.9.8.dylib .
+cp $deps_prefix/lib/libhtmlcxx.3.dylib .
+cp $deps_prefix/lib/libcharsetdetect.dylib .
+cp $deps_prefix/lib/libmimetic.0.dylib .
+cp $deps_prefix/lib/libbfio.1.dylib .
+cp $deps_prefix/lib/libpff.1.dylib .
 cd ..
 
 cd build/tests
 DYLD_FALLBACK_LIBRARY_PATH=.. ctest -j4 -V
 cd ../..
 
-rm -rf build/CMakeFiles build/src build/tests/ build/examples/CMakeFiles build/doc/CMakeFiles
+build_type=$1
+if [ "$build_type" = "--release" ]; then
+	rm -rf build/CMakeFiles build/src build/tests/ build/examples/CMakeFiles build/doc/CMakeFiles
+	rm build/Makefile build/CMakeCache.txt build/cmake_install.cmake build/install_manifest.txt build/examples/cmake_install.cmake build/examples/Makefile build/doc/Makefile build/doc/cmake_install.cmake build/doc/Doxyfile.doxygen
+fi
 
 cd build/
 for i in *.dylib*; do
